Clang 16 (#4)
* Remove unused packages from the container setup.

* Add .bazelrc config.

This instructs Bazel to use the shipped Clang toolchain instead of the
host compiler for building targets.

* Downgrade GCC to 10.4.

GCC 11 introduced __malloc__ attribute parameters so it can only be
compiled with a recent enough GCC. Current Clang (16.0.0) does not
support the __malloc__ params.

* Upgrade Clang to 16.0.0.

Unfortunately this means downgrading the Ubuntu image to 18.04, since
the latest Clang is not built against Ubuntu 20. It also means that
builds now no longer pass on a recent enough host, due to Clang
requiring tinfo 5 (recent hosts ship with version 6).

Ideally we should build our own Clang, but it is very time consuming.